lambda/requirements.txt
Pillow==10.1.0
boto3==1.29.7
